+++ sickrage/providers/torrent/elitetorrent.py
@@ -78,7 +78,7 @@ class EliteTorrentProvider(TorrentProvider):
                 if mode != 'RSS':
                     sickrage.app.log.debug("Search string: {0}".format(search_string))
 
-                search_string = re.sub(r'S0*(\d*)E(\d*)', r'\1x\2', search_string)
+                search_string = re.sub(r'S0*(\d*)E?(\d*)', r'\1x\2', search_string)
                 search_params['buscar'] = search_string.strip() if mode != 'RSS' else ''
 
                 try: